Prevent BuildController from running builds
As opposed to preventing the actual creation of builds, which is hard, this
approach has proved surprisingly easy.

The downside is that builds can still be created. They will fail instantly,
transitioning to error state, but they will still be there. Future efforts
should focus on the elimination of the ability to create builds based on the
OpenshiftEnabled conditional.